MirroringTable: extract batch methods to batcher class
