+++ bigtable-hbase-mirroring-client-1.x-parent/bigtable-hbase-mirroring-client-1.x/src/main/java/com/google/cloud/bigtable/mirroring/hbase1_x/utils/BatchHelpers.java
@@ -19,14 +19,24 @@ import static com.google.cloud.bigtable.mirroring.hbase1_x.utils.OperationUtils.
 
 import com.google.cloud.bigtable.mirroring.hbase1_x.MirroringOperationException;
 import com.google.cloud.bigtable.mirroring.hbase1_x.MirroringOperationException.ExceptionDetails;
+import com.google.cloud.bigtable.mirroring.hbase1_x.WriteOperationFutureCallback;
+import com.google.cloud.bigtable.mirroring.hbase1_x.utils.OperationUtils.RewrittenIncrementAndAppendIndicesInfo;
+import com.google.cloud.bigtable.mirroring.hbase1_x.utils.flowcontrol.FlowController;
+import com.google.cloud.bigtable.mirroring.hbase1_x.utils.flowcontrol.RequestResourcesDescription;
+import com.google.cloud.bigtable.mirroring.hbase1_x.utils.flowcontrol.WriteOperationInfo;
 import com.google.cloud.bigtable.mirroring.hbase1_x.utils.mirroringmetrics.MirroringSpanConstants.HBaseOperation;
 import com.google.cloud.bigtable.mirroring.hbase1_x.utils.mirroringmetrics.MirroringTracer;
 import com.google.cloud.bigtable.mirroring.hbase1_x.verification.MismatchDetector;
+import com.google.common.base.Function;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Predicate;
+import com.google.common.base.Supplier;
 import com.google.common.util.concurrent.FutureCallback;
+import com.google.common.util.concurrent.ListenableFuture;
+import com.google.common.util.concurrent.MoreExecutors;
 import io.opencensus.common.Scope;
 import java.io.IOException;
+import java.io.InterruptedIOException;
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.IdentityHashMap;
@@ -42,6 +52,7 @@ import org.apache.hadoop.hbase.client.RetriesExhaustedWithDetailsException;
 import org.apache.hadoop.hbase.client.Row;
 import org.apache.hadoop.hbase.client.RowMutations;
 import org.apache.hadoop.hbase.client.Table;
+import org.apache.hadoop.hbase.client.coprocessor.Batch.Callback;
 import org.checkerframework.checker.nullness.compatqual.NullableDecl;
 
 public class BatchHelpers {
@@ -291,6 +302,15 @@ public class BatchHelpers {
     }
   }
 
+  /**
+   * Analyses results of two batch operations run concurrently and gathers results into {@code
+   * outputResult} array.
+   *
+   * <p>If there were any failed operations in one of the batches a {@link
+   * RetriesExhaustedWithDetailsException} is thrown. Exceptions stored inside the thrown exception
+   * and in {@code outputResults} are marked with {@link MirroringOperationException} denoting
+   * whether operation have failed on primary, on secondary or on both databases.
+   */
   public static void reconcileBatchResultsConcurrent(
       Object[] outputResults,
       BatchData primaryBatchData,
@@ -355,6 +375,15 @@ public class BatchHelpers {
     }
   }
 
+  /**
+   * Analyses results of two batch operations run sequentially (failed primary operation were not
+   * mirrored to secondary) and gathers results and errors in {@code outputResults} array.
+   *
+   * <p>If there were any failed operations in one of the batches a {@link
+   * RetriesExhaustedWithDetailsException} is thrown. Exceptions stored inside the thrown exception
+   * and in {@code outputResults} are marked with {@link MirroringOperationException} denoting
+   * whether operation have failed on primary or on secondary database.
+   */
   public static void reconcileBatchResultsSequential(
       Object[] outputResults,
       BatchData primaryBatchData,
